+++ endpoint/http/src/main/java/de/fraunhofer/iosb/ilt/faaast/service/endpoint/http/request/mapper/DeleteConceptDescriptionByIdRequestMapper.java
@@ -41,7 +41,7 @@ public class DeleteConceptDescriptionByIdRequestMapper extends AbstractRequestMa
     @Override
     public Request doParse(HttpRequest httpRequest, Map<String, String> urlParameters) {
         return DeleteConceptDescriptionByIdRequest.builder()
-                .id(IdentifierHelper.parseIdentifier(EncodingHelper.base64Decode(urlParameters.get(CONCEPT_ID))))
+                .id(IdentifierHelper.parseIdentifier(EncodingHelper.base64UrlDecode(urlParameters.get(CONCEPT_ID))))
                 .build();
     }
 }